+++ project/apps/init/src/app/main-loop.ts
@@ -1,5 +1,6 @@
 import { ChildProcessWithoutNullStreams } from 'child_process';
 import { getConfig } from '@tx/config';
+import { logger } from '@tx/logger';
 
 import { exitAllProcesses } from './exit-process';
 import { renewalProcess } from './renewal-process';
@@ -10,13 +11,18 @@ import { setIntervalWithoutDelay } from '@tx/util';
  * @param nginx nginx process stream
  */
 export const startMainLoop = (nginx: ChildProcessWithoutNullStreams) => {
-	console.log('[init] Starting main loop...');
+	logger.info('Starting main loop...');
 
 	const timer = setIntervalWithoutDelay(async () => {
-		const status = await renewalProcess();
+		const status = renewalProcess();
 		if (status > 0) {
-			timer.unref();
+			if (timer) timer.unref();
 			exitAllProcesses('renewal failed', nginx);
 		}
+		logger.info(
+			`Start next renewal process in ${
+				getConfig().letsEncrypt.renewalTimer
+			} seconds...`
+		);
 	}, getConfig().letsEncrypt.renewalTimer * 1000);
 };