+++ project/apps/init/src/app/renewal-process.ts
@@ -1,17 +1,19 @@
+import { spawnSync } from 'child_process';
+
 import { requestCertificate } from '@tx/certbot';
 import { getConfig } from '@tx/config';
 import { enableDomain, getDomains } from '@tx/domain';
-import { execute } from '@tx/util';
+import { logger } from '@tx/logger';
 
 /**
  * Main certificate renewal process
  * @returns code 0 when process successful
  */
-export const renewalProcess = async (): Promise<number> => {
-	console.log('Starting certificate renewal process');
+export const renewalProcess = (): number => {
+	logger.info('Starting certificate renewal process');
 
 	if (!process.env.CERTBOT_EMAIL) {
-		console.error(
+		logger.error(
 			'CERTBOT_EMAIL environment variable undefined; certbot will do nothing!'
 		);
 		return 1;
@@ -21,10 +23,10 @@ export const renewalProcess = async (): Promise<number> => {
 
 	const domains = getDomains();
 	if (domains.length) {
-		console.log(`Found ${domains.length} domains to request certificates for`);
+		logger.info(`Found ${domains.length} domains to request certificates for`);
 		domains.forEach(async (domain) => {
 			// TODO: Drop extra domains feature and look for server_name in config file instead
-			const status = await requestCertificate(domain);
+			const status = requestCertificate(domain);
 			if (!status) {
 				exitCode = 2;
 			}
@@ -32,22 +34,22 @@ export const renewalProcess = async (): Promise<number> => {
 		});
 
 		// Let nginx reload its configuration
-		const execStatus = await execute('nginx -s reload');
-		if (execStatus.stderr) {
-			console.error('ERROR: nginx reload failed');
-			console.error(execStatus.stderr);
+		const status = spawnSync('nginx', ['-s', 'reload']);
+		if (status.error) {
+			logger.error('nginx reload failed');
+			logger.error(status.error.message);
 			exitCode = 3;
 		}
 	} else {
-		console.warn('Found no domains to request certificates for');
-		console.log(
+		logger.warn('Found no domains to request certificates for');
+		logger.warn(
 			`Make sure configurations are saved to ${
 				getConfig().nginx.userConfigPath
 			} with valid content - domain in filename and certificate name inside file must match!`
 		);
 	}
 
-	console.log(`End certificate renewal process with code ${exitCode}`);
+	logger.info(`End certificate renewal process with code ${exitCode}`);
 
 	return exitCode;
 };