+++ project/libs/certbot/src/lib/certbot.ts
@@ -1,7 +1,9 @@
+import { spawnSync, SpawnSyncReturns } from 'child_process';
+
 import { getConfig, getLetsEncryptServer } from '@tx/config';
 import { getEnv } from '@tx/environment';
+import { logger } from '@tx/logger';
 import { getStore } from '@tx/store';
-import { execute } from '@tx/util';
 
 /**
  * Request a certificate for the primary domain and optional domains.
@@ -9,16 +11,16 @@ import { execute } from '@tx/util';
  *
  * @param primaryDomain Primary domain for certificate
  * @param optionalDomains Domain variants to add to the same certificate
+ * @returns `true` when request was successful
  *
  * @todo
  * #TODO: Implement support for `optionalDomains`
- * #TODO: Implement support for `--dry-run`
  */
-export const requestCertificate = async (
+export const requestCertificate = (
 	primaryDomain: string,
 	optionalDomains?: string[]
-): Promise<boolean> => {
-	console.log(`Request certificate for primary domain ${primaryDomain}`);
+): boolean => {
+	logger.info(`Request certificate for primary domain ${primaryDomain}...`);
 
 	// Optional domains are provided with `-d` flag before each domain
 	optionalDomains = optionalDomains ?? [];
@@ -27,28 +29,53 @@ export const requestCertificate = async (
 	}${optionalDomains.join(' -d ')}`;
 
 	const forceRenewal = getStore().forceRenew ? '--force-renewal' : '';
+	const dryRun = getEnv().DRY_RUN === 'Y' ? '--dry-run' : '';
+	const isolated = getEnv().ISOLATED === 'Y';
 
 	// Create certbot command
 	const config = getConfig().letsEncrypt;
-	const command = `
-      certbot certonly \
-        --agree-tos --keep -n --text \
-        -a webroot --webroot-path=${config.webRoot} \
-        --rsa-key-size ${config.rsaKeySize} \
-        --preferred-challenges http-01 \
-        --email ${getEnv().CERTBOT_EMAIL} \
-        --server ${getLetsEncryptServer()} \
-        --cert-name ${primaryDomain} \
-        ${includeDomainArg} \
-        ${forceRenewal} \
-        --debug`;
+	const command = 'certbot';
+	const args = [
+		'certonly',
+		'--agree-tos',
+		'--keep',
+		'-n',
+		'--text',
+		'-a',
+		'webroot',
+		`--webroot-path=${config.webRoot}`,
+		'--rsa-key-size',
+		`${config.rsaKeySize}`,
+		'--preferred-challenges',
+		'http-01',
+		'--email',
+		`${getEnv().CERTBOT_EMAIL}`,
+		'--server',
+		`${getLetsEncryptServer()}`,
+		'--cert-name',
+		`${primaryDomain}`,
+		`${includeDomainArg}`,
+		`${forceRenewal}`,
+		`${dryRun}`,
+		'--debug'
+	];
 
-	// Execute command and hence request the certificate
-	const status = await execute(command);
-	if (status.stderr) {
-		console.error(status.stderr);
+	let status: SpawnSyncReturns<string>;
+	if (!isolated) {
+		// Spawn command syncron and hence request the certificate
+		status = spawnSync(command, args);
+	} else {
+		logger.info('Running in isolated mode, no request will be made!');
+		logger.info('certbot request command:');
+		logger.info(command);
+		status = { error: null } as SpawnSyncReturns<string>;
+	}
+	if (status.error) {
+		logger.error(`Failed with message '${status.error.message}'`);
+	} else {
+		logger.info('Renewal done');
 	}
 
-	// Return false when we have something in stderr
-	return status.stderr === '';
+	// Return false when we have an error
+	return !status.error;
 };