+++ project/libs/domain/src/lib/domain.spec.ts
@@ -1,7 +1,9 @@
 import * as fs from 'fs';
 import { basename, dirname, join } from 'path';
 import * as fsMock from 'mock-fs';
+
 import { Config } from '@tx/config';
+import { logger } from '@tx/logger';
 
 import {
 	disablePendingDomains,
@@ -13,22 +15,34 @@ import {
 jest.mock('@tx/config', () => ({
 	getConfig: jest.fn().mockReturnValue({
 		cert: { domainPath: '', privatePem: 'privkey.pem' },
-		nginx: { configPath: 'nginx_config.d', userConfigPath: 'config.d' }
+		nginx: { configPath: 'config.d', userConfigPath: 'user.config.d' }
 	} as Config)
 }));
 
 // Mock a filesystem
 fsMock({
-	'config.d': {
+	'user.config.d': {
 		'domain.com.conf': '',
 		'site.com.conf': ''
 	},
-	'nginx_config.d': fsMock.load(join(dirname(__filename), 'test-files'), {
+	'config.d': fsMock.load(join(dirname(__filename), 'test-files', 'config.d'), {
 		lazy: false
-	})
+	}),
+	'secure.d': fsMock.load(join(dirname(__filename), 'test-files', 'secure.d'), {
+		lazy: false
+	}),
+	ssl: fsMock.load(join(dirname(__filename), 'test-files', 'ssl'), {
+		lazy: false
+	}),
+	letsencrypt: fsMock.load(
+		join(dirname(__filename), 'test-files', 'letsencrypt'),
+		{
+			lazy: false
+		}
+	)
 });
 
-console.log = jest.fn();
+logger.info = jest.fn();
 
 describe('domain', () => {
 	afterAll(() => fsMock.restore());
@@ -39,13 +53,17 @@ describe('domain', () => {
 
 	it('should transfer user config files', () => {
 		transferUserConfig();
-		expect(fs.existsSync('nginx_config.d/valid-my-site.com.conf')).toBeTruthy();
-		expect(fs.existsSync('nginx_config.d/domain.com.conf')).toBeTruthy();
-		expect(fs.existsSync('nginx_config.d/site.com.conf')).toBeTruthy();
+		expect(fs.existsSync('config.d/valid-my-site.com.conf')).toBeTruthy();
+		expect(fs.existsSync('config.d/domain.com.conf')).toBeTruthy();
+		expect(fs.existsSync('config.d/site.com.conf')).toBeTruthy();
 	});
 
-	it('should only find valid sites', () => {
-		expect(getDomains()).toEqual(['valid-keys-site.com', 'valid-my-site.com']);
+	it('should only find valid domain names matching key file', () => {
+		expect(getDomains()).toEqual([
+			'valid-keys-site.com',
+			'valid-keys-site2.com',
+			'valid-my-site.com'
+		]);
 	});
 
 	it('should not disable valid-keys-site.com', () => {
@@ -88,4 +106,8 @@ describe('domain', () => {
 		enableDomain('valid-keys-site.com');
 		expect(fs.rename).toHaveBeenCalledTimes(0);
 	});
+
+	it.todo('should disable domain that are missing dhparam.pem');
+	it.todo('should not enale domain that are missing dhparam.pem');
+	it.todo('should emable pending domain that have dhparam.pem');
 });