+++ project/libs/nginx/src/index.ts
@@ -1,1 +1,2 @@
+export * from './lib/diffie-hellman';
 export * from './lib/nginx';