+++ CHANGELOG.md
@@ -1,9 +1,35 @@
-# 0.2
+# 0.3 (2021-01-24)
+
+Features
+
+- Use DRY_RUN environment to apply certbot `--dry-run` flag
+- Use ISOLATED environment to run without real certbot requests
+- Provide script to create self signed Let's Encrypt certs for isoleted test
+- Use Winston for logging
+- Use Diffie-Hellman parameters file
+
+Enhancement
+
+- Use syncron child processes only to stop using async await decorators
+- Add more tests to `certbot` lib
+- Improve ssl security to get A+ by SSL Labs
+- Remove support for TLS V1.1, it's soon deprecated
+
+Bug fix
+
+- Second domain request failed due to certbot lock file
+- Various fixes...
+
+# 0.2 (2021-01-15)
+
+Enhancement
+
+- Add optional environment variables to readme
 
 Bug fix
 
 - Renewal failed for single domain request
 
-# 0.1
+# 0.1 (2021-01-13)
 
 - Major rewrite to Node and TypeScript